backend

+++ BlogApplication2/src/main/java/blog/example/BlogApplication2/Service/PostService.java
@@ -1,10 +1,11 @@
 package blog.example.BlogApplication2.Service;
-import blog.example.BlogApplication2.Model.BlogPostHistory;
-import blog.example.BlogApplication2.Model.Comment;
-import blog.example.BlogApplication2.Repository.BlogPostHistoryRepository;
+import blog.example.BlogApplication2.Model.ChangeType;
+import blog.example.BlogApplication2.Model.History;
+import blog.example.BlogApplication2.Repository.HistoryRepository;
 import blog.example.BlogApplication2.Repository.LikeRepository;
 import blog.example.BlogApplication2.Repository.PostRepository;
 import blog.example.BlogApplication2.Model.Blogpost;
+import jakarta.persistence.PreRemove;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,7 +16,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class PostService {
@@ -24,17 +24,17 @@ public class PostService {
     private LikeRepository likeRepository;
 
     private ImageUploadService imageUploadService;
-    private BlogPostHistoryRepository blogPostHistoryRepository;
+    private HistoryRepository historyRepository;
 
     @Autowired
     public PostService(PostRepository postRepository,
                        LikeRepository likeRepository,
                        ImageUploadService imageUploadService,
-                       BlogPostHistoryRepository blogPostHistoryRepository) {
+                       HistoryRepository historyRepository) {
         this.postRepository = postRepository;
         this.likeRepository = likeRepository;
         this.imageUploadService=imageUploadService;
-        this.blogPostHistoryRepository=blogPostHistoryRepository;
+        this.historyRepository=historyRepository;
 
     }
 
@@ -51,16 +51,24 @@ public class PostService {
         return postRepository.save(blogpost);
     }
 
-    @Transactional
-    public boolean deleteBlogpost(Integer postid) {
+    @PreRemove
+    public void deleteBlogpost(Integer postid) throws Exception {
         try {
-
-            likeRepository.deleteByPostId(postid);
-            postRepository.deleteById(postid);
-            return true;
-
+            Blogpost deletedPost=postRepository.findById(postid).orElse(null);
+            if(deletedPost!=null) {
+                History history = new History();
+                history.setOldcontent(deletedPost.getContent());
+                history.setOldtitle(deletedPost.getTitle());
+                history.setOldimage(deletedPost.getImage());
+                history.setTime(new Date());
+                history.setChangetype(ChangeType.DELETE);
+                history.setBlogpost(deletedPost);
+                historyRepository.save(history);
+                likeRepository.deleteByPostId(postid);
+                postRepository.deleteById(postid);
+            }
         } catch (Exception e) {
-            return false;
+            e.printStackTrace();
         }
 
     }
@@ -80,18 +88,18 @@ public class PostService {
         Blogpost existingPost = postRepository.findById(postid).orElse(null);
 
         if (existingPost != null) {
-            BlogPostHistory history = new BlogPostHistory();
-            history.setNewcontent(existingPost.getContent()); 
-            history.setNewtitle(existingPost.getTitle());
-            history.setNewimage(existingPost.getImage());
-            history.setBlogpost(existingPost);
-            history.setTime(new Date());
-            blogPostHistoryRepository.save(history);
-            existingPost.setContent(updatedPost.getContent());
-            existingPost.setTitle(updatedPost.getTitle());
-            existingPost.setImage(updatedPost.getImage());
-
-            return postRepository.save(existingPost);
+            History history = new History();
+           history.setOldcontent(existingPost.getContent());
+           history.setOldtitle(existingPost.getTitle());
+           history.setBlogpost(existingPost);
+           history.setOldimage(existingPost.getImage());
+           history.setChangetype(ChangeType.EDIT);
+           history.setTime(new Date());
+           historyRepository.save(history);
+           existingPost.setTitle(updatedPost.getTitle());
+           existingPost.setContent(updatedPost.getContent());
+           existingPost.setImage(updatedPost.getImage());
+           return postRepository.save(existingPost);
         }
         return null;
     }
